fix: create types directory for component adapter props
