+++ src/components/Common/Fields/ComboBoxField/ComboBoxField.tsx
@@ -1,8 +1,8 @@
 import { useField, type UseFieldProps } from '@/components/Common/Fields/hooks/useField'
 import { ComboBox } from '@/components/Common/UI/ComboBox/ComboBox'
-
+import type { ComboBoxProps } from '@/types/UI/ComboBox'
 interface ComboBoxFieldProps
-  extends Omit<React.ComponentProps<typeof ComboBox>, 'name' | 'onChange' | 'onBlur'>,
+  extends Omit<ComboBoxProps, 'name' | 'onChange' | 'onBlur'>,
     UseFieldProps<string> {}
 
 export const ComboBoxField: React.FC<ComboBoxFieldProps> = ({
