+++ src/components/Common/Fields/SelectField/SelectField.tsx
@@ -1,6 +1,6 @@
 import { useField, type UseFieldProps } from '@/components/Common/Fields/hooks/useField'
-import { Select, type SelectProps } from '@/components/Common/UI/Select'
-
+import { Select } from '@/components/Common/UI/Select'
+import type { SelectProps } from '@/types/UI/Select'
 interface SelectFieldProps
   extends Omit<SelectProps, 'name' | 'value' | 'onChange' | 'onBlur'>,
     UseFieldProps<string> {}
