+++ src/components/Common/UI/Checkbox/index.ts
@@ -1,2 +1,1 @@
 export { Checkbox } from './Checkbox'
-export type { CheckboxProps } from './Checkbox'
