+++ src/components/Common/UI/FieldLayout/index.ts
@@ -1,6 +1,1 @@
 export { FieldLayout } from './FieldLayout'
-export type {
-  SharedFieldLayoutProps,
-  InternalFieldLayoutProps,
-  FieldLayoutProps,
-} from './FieldLayout'
