+++ src/components/Common/UI/HorizontalFieldLayout/index.ts
@@ -1,5 +1,1 @@
 export { HorizontalFieldLayout } from './HorizontalFieldLayout'
-export type {
-  SharedHorizontalFieldLayoutProps,
-  HorizontalFieldLayoutProps,
-} from './HorizontalFieldLayout'
