+++ src/components/Common/UI/Radio/index.ts
@@ -1,2 +1,1 @@
 export { Radio } from './Radio'
-export type { RadioProps } from './Radio'
