+++ src/components/Common/UI/Select/index.ts
@@ -1,1 +1,1 @@
-export { Select, type SelectProps } from './Select'
+export { Select } from './Select'
