+++ src/components/Common/UI/TextInput/index.ts
@@ -1,1 +1,1 @@
-export { TextInput, type TextInputProps } from './TextInput'
+export { TextInput } from './TextInput'
