+++ src/types/UI/Checkbox.ts
@@ -1,0 +1,14 @@
+import type { InputHTMLAttributes, Ref } from 'react'
+import type { SharedHorizontalFieldLayoutProps } from './HorizontalFieldLayout'
+
+export interface CheckboxProps
+  extends SharedHorizontalFieldLayoutProps,
+    Pick<
+      InputHTMLAttributes<HTMLInputElement>,
+      'name' | 'id' | 'className' | 'onChange' | 'onBlur' | 'checked' | 'value'
+    > {
+  inputRef?: Ref<HTMLInputElement>
+  isInvalid?: boolean
+  isDisabled?: boolean
+  inputProps?: InputHTMLAttributes<HTMLInputElement>
+}
