+++ src/types/UI/DatePIcker.ts
@@ -1,0 +1,15 @@
+import type { InputHTMLAttributes, Ref, FocusEvent } from 'react'
+import type { SharedFieldLayoutProps } from './FieldLayout'
+
+export interface DatePickerProps
+  extends SharedFieldLayoutProps,
+    Pick<InputHTMLAttributes<HTMLInputElement>, 'className' | 'id' | 'name'> {
+  inputRef?: Ref<HTMLInputElement>
+  isDisabled?: boolean
+  isInvalid?: boolean
+  onChange?: (value: Date | null) => void
+  onBlur?: (e: FocusEvent) => void
+  label: string
+  value?: Date
+  placeholder?: string
+}
