+++ src/types/UI/HorizontalFieldLayout.ts
@@ -1,0 +1,5 @@
+import type { InternalFieldLayoutProps, SharedFieldLayoutProps } from './FieldLayout'
+
+export type SharedHorizontalFieldLayoutProps = SharedFieldLayoutProps
+
+export type HorizontalFieldLayoutProps = SharedHorizontalFieldLayoutProps & InternalFieldLayoutProps
