+++ src/types/UI/Radio.ts
@@ -1,0 +1,14 @@
+import type { InputHTMLAttributes, Ref } from 'react'
+import type { SharedHorizontalFieldLayoutProps } from './HorizontalFieldLayout'
+
+export interface RadioProps
+  extends SharedHorizontalFieldLayoutProps,
+    Pick<
+      InputHTMLAttributes<HTMLInputElement>,
+      'name' | 'id' | 'className' | 'onChange' | 'onBlur' | 'checked' | 'value'
+    > {
+  inputRef?: Ref<HTMLInputElement>
+  isInvalid?: boolean
+  isDisabled?: boolean
+  inputProps?: InputHTMLAttributes<HTMLInputElement>
+}
