+++ src/types/UI/Select.ts
@@ -1,0 +1,21 @@
+import type { FocusEvent, Ref, SelectHTMLAttributes } from 'react'
+import type { SharedFieldLayoutProps } from './FieldLayout'
+
+export interface SelectItem {
+  value: string
+  label: string
+}
+
+export interface SelectProps
+  extends SharedFieldLayoutProps,
+    Pick<SelectHTMLAttributes<HTMLSelectElement>, 'id' | 'name' | 'className'> {
+  isDisabled?: boolean
+  isInvalid?: boolean
+  label: string
+  onChange: (value: string) => void
+  onBlur?: (e: FocusEvent) => void
+  options: SelectItem[]
+  placeholder?: string
+  value?: string
+  inputRef?: Ref<HTMLButtonElement>
+}
